add a test_utils.rs to the api codebase, with methods to generate temporary test collections, and a method to read a json body to deduplicate logic present in all the tests
