+++ api/src/lib.rs
@@ -6,3 +6,4 @@ pub mod docs;
 pub mod filters;
 pub mod models;
 pub mod routes;
+pub mod test_utils;
