+++ api/tests/test_catalogs.rs
@@ -5,33 +5,8 @@ mod tests {
     use actix_web::{App, test, web};
     use boom_api::db::get_default_db;
     use boom_api::routes;
-    use mongodb::{Database, bson};
-
-    // let's make a function that creates a new catalog for testing
-    // the catalog should have a randomized name to avoid conflicts
-    // and a function to delete the catalog after the test is done
-    async fn create_test_catalog(db: &Database) -> String {
-        let name = format!("test_catalog_{}", uuid::Uuid::new_v4());
-        let collection = db.collection(&name);
-        // add an index to the collection
-        collection
-            .create_index(
-                mongodb::IndexModel::builder()
-                    .keys(mongodb::bson::doc! { "test_field": 1 })
-                    .options(None)
-                    .build(),
-            )
-            .await
-            .unwrap();
-        // insert a dummy document
-        let doc = bson::doc! { "test_field": "test_value" };
-        collection.insert_one(doc).await.unwrap();
-        name
-    }
-
-    async fn delete_test_catalog(db: &Database, name: &str) {
-        db.collection::<bson::Document>(name).drop().await.unwrap();
-    }
+    use boom_api::test_utils::{create_test_catalog, delete_test_catalog, read_json_response};
+    use mongodb::Database;
 
     /// Test GET /catalogs
     #[actix_rt::test]
@@ -47,15 +22,8 @@ mod tests {
 
         let req = test::TestRequest::get().uri("/catalogs").to_request();
         let resp = test::call_service(&app, req).await;
-
         assert_eq!(resp.status(), StatusCode::OK);
-
-        let body = test::read_body(resp).await;
-        let body_str = String::from_utf8_lossy(&body);
-
-        // Parse response body JSON
-        let resp: serde_json::Value =
-            serde_json::from_str(&body_str).expect("failed to parse JSON");
+        let resp = read_json_response(resp).await;
 
         assert!(resp["data"].is_array());
     }
@@ -76,13 +44,7 @@ mod tests {
             .to_request();
         let resp = test::call_service(&app, req).await;
         assert_eq!(resp.status(), StatusCode::OK);
-
-        // Parse response body JSON
-        let body = test::read_body(resp).await;
-        let body_str = String::from_utf8_lossy(&body);
-
-        let resp: serde_json::Value =
-            serde_json::from_str(&body_str).expect("failed to parse JSON");
+        let resp = read_json_response(resp).await;
 
         assert!(resp["data"].is_array());
 
@@ -105,11 +67,8 @@ mod tests {
             .to_request();
         let resp = test::call_service(&app, req).await;
         assert_eq!(resp.status(), StatusCode::OK);
-        // Parse response body JSON
-        let body = test::read_body(resp).await;
-        let body_str = String::from_utf8_lossy(&body);
-        let resp: serde_json::Value =
-            serde_json::from_str(&body_str).expect("failed to parse JSON");
+        let resp = read_json_response(resp).await;
+
         assert!(resp["data"].is_array());
         assert_eq!(resp["data"].as_array().unwrap().len(), 1);
         // Clean up test catalog
